fixes based on review comments

+++ apis/config/configset_types.go
@@ -31,7 +31,7 @@ type ConfigSetSpec struct {
 	// will follow
 	Lifecycle *Lifecycle `json:"lifecycle,omitempty" protobuf:"bytes,2,opt,name=lifecycle"`
 	// Priority defines the priority of this config
-	Priority int64 `json:"priority,omitempty" protobuf:"varint,3,opt,name=priority"`
+	Priority int32 `json:"priority,omitempty" protobuf:"varint,3,opt,name=priority"`
 	// Revertive defines if this CR is enabled for revertive or non revertve operation
 	Revertive *bool `json:"revertive,omitempty" protobuf:"varint,4,opt,name=revertive"`
 	// Config defines the configuration to be applied to a target device
